Implement base dashboard layout for consistent UI structure
