+++ layouts/pacientes-dashboard.vue
@@ -1,7 +1,13 @@
 <template>
-  <div>
+  <div class="layout-container">
     <PacientesNav />
     <slot />
   </div>
 </template>
 <style src="@/assets/styles/vitalink.scss" />
+
+<style lang="scss" scoped>
+.layout-container {
+  overflow-x: hidden;
+}
+</style>
